Runscripts and tools for Gaea CI
These are test scripts for running the CI tests on Gaea for MOM6.
They're unlikely to ever be integrated into the repos, most likely for
personal testing.

The Makefile is designed to be used in the MOM6-examples repo, and
currently builds three of the required executables.  It doesn't yet
distinguish between symmetric or static memory.

The Python script is a work in progress to generate the list of
regression tests.  There is also a code block which can launch the tests
in an interactive Slurm session.  But there is no checking that the job
completed, nor is there any comparison of output against the reference.

Much work to do...